Allow only gerrit-verifiers to upload patch sets to changes of other users
There are some spammers that upload new patch sets to changes that
belong to other users (e.g. [1]). To prevent that this happens again
we grant the 'Add Patch Set' permission [2] only to the gerrit-verifiers
group (and block it for Anonymous-Users since on the All-Projects project
this permission is granted for Registered-Users).

This means only users of the gerrit-verifiers can now upload new patch sets
to changes that they do not own. Everyone is still allowed to upload patch
sets to own changes since this is always allowed regardless of the
'Add Patch Set' permission.

[1] https://gerrit-review.googlesource.com/c/plugins/javamelody/+/192431/10..11//COMMIT_MSG
[2] https://gerrit-review.googlesource.com/Documentation/access-control.html#category_add_patch_set

Change-Id: I99263899fa3a82cbefbc20550b2e07dfd9414184